move login and agree_tos into login package
