+++ src/pkg/mcp/tools/login.go
@@ -5,6 +5,7 @@ import (
 	"strconv"
 
 	"github.com/DefangLabs/defang/src/pkg/cli"
+	"github.com/DefangLabs/defang/src/pkg/login"
 	"github.com/DefangLabs/defang/src/pkg/term"
 	"github.com/DefangLabs/defang/src/pkg/track"
 	"github.com/mark3labs/mcp-go/mcp"
@@ -33,7 +34,7 @@ func setupLoginTool(s *server.MCPServer, cluster string, authPort int) {
 				return mcp.NewToolResultText("Please open this URL in your browser: http://127.0.0.1:" + strconv.Itoa(authPort) + " to login"), nil
 			}
 			term.Debug("Function invoked: cli.InteractiveLoginPrompt")
-			err = cli.InteractiveLoginMCP(ctx, client, cluster)
+			err = login.InteractiveLoginMCP(ctx, client, cluster)
 			if err != nil {
 				return mcp.NewToolResultErrorFromErr("Failed to login", err), nil
 			}
